Improve BAM record building and memory efficiency
Instead of allocating individual BAM records using bam_init,
make processRecords() allocate two blocks of memory - one big enough
to hold all the bam1_t structs it will produce, and one big enough
for all of the variable-length parts.  The bam records themselves
are created in small groups so that most of the data needed will
be in cache.  In particular, we try to access the data in the bcl
bases and quals arrays efficiently.  They are also made by writing
directly into the bam1_t structure.  This is terrible from an
encapsulation point of view, but much quicker than trying to
use interfaces like bam_aux_append.

Some work that was being repeated (e.g. calls to getBclFileArray;
working out the first part of the read name) is lifted up out
of the inner loops.  Some work on grouping barcode tags is done
up front so that they can be written more efficiently.

Tidy up functions and variables that are no longer needed.