Simplify opening BCL files
The htslib thread pool normally attempts to return results in
the same order as jobs were submitted.  In the case of opening
the BCL files that's not really necessary.  Instead we can
use an "input only" thread pool process to open the BCL files
and get bcl_thread() to slot the bclfile_t pointer into the
right place in the output array.

Makes openBclFiles much simpler as it no longer needs to pull
results back from the thread pool queue.  It also allows the jobs
to run more efficiently as we avoid occasional delays while we
want for the "next" result to come back.