Cache BCL files for NovaSeq
NovaSeq CBCL files contain data for all tiles on a given lane,
surface and cycle.  This means we don't want to close the files
after each tile as we just end up opening the same ones again
(which takes a very long time).  Implement this by keeping the
BCL files in a cache.