Allow use of posix_fadvise on NovaSeq cbcl files
Use posix_fadvise() to tell the filesystem which part of the cbcl
file is going to be read next, and which part is finished with.
Hopefully the data for the next tile will have been loaded into
the buffer cache by the time it's needed.  When the data has been
read, we tell the filesystem we don't need it any more so that it
doesn't hang around for longer than needed.

The exact behaviour can be controlled by defining USE_POSIX_FADVISE
(see the top of src/bclfile.c).