More decode threading
Add processTemplatesNoThreads() and processTemplatesThreads()
which process the data in single- and multi- threaded modes.

Remove sam writing from processTemplate so it can be called
from both the threaded and non-threaded versions.

Input records are grouped into batches by processTemplatesThreads()
for thread pool job decode_job() to work on.  Each batch has its
own copy of barcodeArray and tagHopHash so that metrics can be
updated without having to lock mutexes.  The metrics in the
copies are added up at the end of the process to give the final
totals.  After a batch has been processed, the job structure
is put onto a fee list so that its copies of barcodeArray,
tagHopHash and the memory used to store bam records can be used
by a future batch.

barcodeHash now stores indexes into barcodeArray instead of
pointers to the items themselves.  This allows a single barcodeHash
to be used for indexing all of the barcodeArray copies.