Enable barcode decoding in i2b. (#113)
* Enable barcode decoding in i2b

Add an interface to decode so that i2b can call into it.  To avoid
a clash with i2b, its opts_t struct is renamed to decode_opts_t.

Add decoding options to i2b.  Options (long only) are copied from
decode.  Decoding is enabled by using the --barcode-file option.

Make addHeader() write a read group header for each barcode
if decoding is enabled instead of a single one for the whole file.

Update processRecordGroup() to make it look up the barcodes and
alter the read group and optionally the read name to say which
tag was found.  The actual decoding is done by decode_tags().
Currently the call data is not shared with bam_write_barcode_tag()
so the barcode bcl files get converted twice.  While not quite
as efficient, it's simpler to keep both separate.

processTile() now recycles processRecordJob_struct structures
using a free-list so that it can reuse barcode metrics structures.

Barcode metrics are accumulated in the same way as for decode and
printed out at the end if the --metrics-file option is used.

In a small test, decoding in i2b saves about 30% CPU compared
to i2b piped into decode.

* Add compare on sequence as last tie-breaker for sorting tag hops

Mainly so files can be easily compared when running tests.

* Add tests for i2b barcode decoding